requirements.txt
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
accelerate>=0.20.0
einops>=0.6.0
pillow>=9.0.0
numpy>=1.24.0
tqdm>=4.65.0
pyyaml>=6.0
jupyter>=1.0.0
ipywidgets>=8.0.0
safetensors>=0.3.1
matplotlib>=3.7.0
wandb>=0.15.0  # optional for logging
